+++ Gruntfile.js
@@ -2,7 +2,6 @@ module.exports = function (grunt) {
 
     grunt.initConfig({
         pkg: grunt.file.readJSON('package.json'),
-        clean: ["assets/*"],
         uglify: {
             calendar: {
                 files: {