+++ integration/BirthdayCalendarQuery.php
@@ -15,7 +15,7 @@
 
 namespace humhub\modules\calendar\integration;
 
-use humhub\modules\calendar\interfaces\AbstractCalendarQuery;
+use humhub\modules\calendar\interfaces\event\AbstractCalendarQuery;
 use humhub\modules\calendar\interfaces\event\FilterNotSupportedException;
 use humhub\modules\content\components\ActiveQueryContent;
 use humhub\modules\space\models\Membership;